+++ crates/affinidi-tdk/common/affinidi-data-integrity/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Affinidi Data Integrity Changelog
 
+## 6th June 2025 Release 0.1.1
+
+* **FEATURE:** Can now verify a JSON Document
+* **FEATURE:** Added example Verification tool for loading signed documents and
+verifying them
+* **FIX:** Serialization of input documents was not correctly handling `@context`
+  * It now correctly handles `@context` fields and places them in the proof
+
 ## 29th May 2025 Release 0.1.0
 
 * Initial release of crate